test_all_imports.py
import os, shutil, tkinter as tk  
from tkinter import ttk, messagebox, simpledialog, filedialog  
from PIL import Image, ImageTk  
from database import db  
import matplotlib  
matplotlib.use('TkAgg')  
import matplotlib.pyplot as plt  
from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg  
import csv, pandas as pd  
from datetime import datetime  
from reportlab.pdfgen import canvas  
from reportlab.lib.pagesizes import A4  
from reportlab.lib import colors  
print('All imports done')  
